requirement.txt
streamlit==1.12.2
pandas==1.4.4
numpy==1.23.2
seaborn
matplotlib
scikit-learn
xgboost
firebase-admin
python-dotenv
streamlit